Fix implicit 64-to-32 bit integer narrowing
Summary:
These fire when a 64-bit integer (e.g., `size_t`, `int64_t`) is implicitly converted to a 32-bit type (e.g., `int`, `uint32_t`), which can cause data loss.

Fixes applied:
- Widened variable types to match source (e.g., `int` → `size_t`)
- Added explicit `static_cast` where narrowing is intentional and safe

Differential Revision: D97079855

+++ momentum/character/parameter_transform.cpp
@@ -266,7 +266,9 @@ ParameterTransformT<T> mapParameterTransformJoints(
 
       triplets.push_back(
           Eigen::Triplet<float>(
-              mappedJoint * kParametersPerJoint + jOffset, static_cast<int>(it.col()), it.value()));
+              static_cast<int>(mappedJoint * kParametersPerJoint + jOffset),
+              static_cast<int>(it.col()),
+              it.value()));
 
       // enable joint channels
       mappedTransform.activeJointParams[mappedJoint * kParametersPerJoint + jOffset] = true;
@@ -364,7 +366,7 @@ std::tuple<ParameterTransformT<T>, ParameterLimits> subsetParameterTransform(
     if (paramOld >= 0) {
       auto paramNew = oldParamToNewParam[paramOld];
       if (paramNew != kInvalidIndex) {
-        paramTransformNew.blendShapeParameters(iBlendShape) = paramNew;
+        paramTransformNew.blendShapeParameters(iBlendShape) = static_cast<int>(paramNew);
       }
     }
   }
@@ -378,7 +380,7 @@ std::tuple<ParameterTransformT<T>, ParameterLimits> subsetParameterTransform(
     if (paramOld >= 0) {
       auto paramNew = oldParamToNewParam[paramOld];
       if (paramNew != kInvalidIndex) {
-        paramTransformNew.faceExpressionParameters(iBlendShape) = paramNew;
+        paramTransformNew.faceExpressionParameters(iBlendShape) = static_cast<int>(paramNew);
       }
     }
   }
